Extract `useUserEmailAddresses`
